CondoMember CRUD

+++ ProjectCondoManagement/Controllers/CondoMembersController.cs
@@ -1,4 +1,5 @@
-﻿using ClassLibrary.DtoModels;
+﻿using ClassLibrary;
+using ClassLibrary.DtoModels;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -25,19 +26,26 @@ namespace ProjectCondoManagement.Controllers
         private readonly DataContextCondos _context;
         private readonly ICondoMemberRepository _condoMemberRepository;
         private readonly IConverterHelper _converterHelper;
+        private readonly IUserHelper _userHelper;
 
-        public CondoMembersController(DataContextCondos context, ICondoMemberRepository condoMemberRepository, IConverterHelper converterHelper)
+        public CondoMembersController(DataContextCondos context, ICondoMemberRepository condoMemberRepository, IConverterHelper converterHelper, IUserHelper userHelper)
         {
             _context = context;
             _condoMemberRepository = condoMemberRepository;
             _converterHelper = converterHelper;
+            _userHelper = userHelper;
         }
 
         // GET: api/CondoMembers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<CondoMember>>> GetCondoMembers()
         {
-            return await _condoMemberRepository.GetAll(_context).ToListAsync();
+            var condoMembers = await _condoMemberRepository.GetAll(_context).ToListAsync();
+
+            await _condoMemberRepository.LinkImages(condoMembers); // Link images to condo members
+            
+            return condoMembers;
+
         }
 
         // GET: api/CondoMembers/5
@@ -45,6 +53,9 @@ namespace ProjectCondoManagement.Controllers
         public async Task<ActionResult<CondoMember>> GetCondoMember(int id)
         {
             var condoMember = await _condoMemberRepository.GetByIdAsync(id, _context);
+
+            var condoMembers = new List<CondoMember> { condoMember };// Create a list with the single condo member for linking images
+            await _condoMemberRepository.LinkImages(condoMembers); // Link images to the single condo member
 
             if (condoMember == null)
             {
@@ -79,7 +90,7 @@ namespace ProjectCondoManagement.Controllers
             {
                 await _condoMemberRepository.UpdateAsync(condoMember, _context);
 
-                return Ok();
+                return Ok(new Response { IsSuccess = true, Message = "Condo member updated successfully." });
             }
             catch (Exception ex)
             {
@@ -111,14 +122,9 @@ namespace ProjectCondoManagement.Controllers
 
                 await _condoMemberRepository.CreateAsync(condoMember, _context);
 
-                return Ok(); 
+                return Ok(new Response { IsSuccess = true, Message = "Condo member created successfully." });
             }
-            //catch (Exception ex)
-            //{
-            //    //return StatusCode(StatusCodes.Status500InternalServerError, "An error occurred while creating the condo member.");
-            //    return StatusCode(StatusCodes.Status500InternalServerError,
-            //        $"An error occurred: {ex.Message}");
-            //}
+
             catch (Exception ex)
             {
                 var errorMessage = ex.InnerException?.Message ?? ex.Message;
@@ -130,7 +136,7 @@ namespace ProjectCondoManagement.Controllers
 
         // DELETE: api/CondoMembers/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteCondoMember(int id)
+        public async Task<IActionResult> DeleteCondoMemberAndDeactivateUser(int id)
         {
             var condoMember = await _condoMemberRepository.GetByIdAsync(id, _context);
             if (condoMember == null)
@@ -138,9 +144,21 @@ namespace ProjectCondoManagement.Controllers
                 return NotFound();
             }
 
+            var user = await _userHelper.GetUserByEmailAsync(condoMember.Email);
+            if (user == null)
+            {
+                return NotFound();
+            }
+
             await _condoMemberRepository.DeleteAsync(condoMember, _context);
 
-            return NoContent();
+            var result = await _userHelper.DeactivateUserAsync(user);
+            if (!result.IsSuccess)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, $"An error occurred while deactivating the user.{result.Message}");
+            }
+
+            return Ok(new Response { IsSuccess = true, Message = "Condo member deleted and user deactivated successfully." });
         }
